storageinventory: Refactor ceph key generation
ceph-ansible 3.1 (and newer) refactored the handling of ceph keys
by leveraging the new ceph-key ansible module, which escapes our
current '$(ceph-authtool ...)' string, ending up as a literal string
instead of evaluating the command and catching its output.

To support the new ceph-ansible 3.1, we refactor the ceph key
generation in storageinventory.py by explicitly running ceph-authtool
with subprocess.check_output and passing it down via a variable.
This change is backwards compatible with the current ceph-ansible 3.0.

Signed-off-by: Alexandru Antone <Alexandru.Antone@enea.com>
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>
Change-Id: I1f0d73399429ef40f414efeb6931db111eba0faf